fixed conflicts, and rename components
