+++ src/components/Popup/Button.tsx
@@ -7,7 +7,7 @@ interface IProps {
 
 const Button: FC<IProps> = ({ text, backgroundColor = "black" }) => (
   <div className="flex justify-center">
-    <button className={`bg-${backgroundColor} h-10 w-5/6 rounded-lg  `}>
+    <button className={`bg-${backgroundColor} h-10 w-5/6 rounded-lg`}>
       <div className="text-white font-lato text-base">{text}</div>
     </button>
   </div>
