+++ src/stories/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
-import KreditPopup from "../components/Popup/Storybook/KreditPopup";
+import CreditPopup from "../components/Popup/Storybook/CreditPopup";
 import ErrorPopup from "../components/Popup/Storybook/ErrorPopup";
 import RegistrationPopup from "../components/Popup/Storybook/RegistrationPopup";
 import CancelReservationPopup from "../components/Popup/Storybook/CancelReservationPopup";
@@ -8,7 +8,7 @@ import "../style/index.css";
 import "../style/fonts.css";
 
 storiesOf("Popups", module)
-  .add("KreditPopup", () => <KreditPopup />)
+  .add("CreditPopup", () => <CreditPopup />)
   .add("RegistrationModal", () => <RegistrationPopup />)
   .add("CancelReservation", () => <CancelReservationPopup />)
   .add("ErrorPopup", () => <ErrorPopup />);
